+++ examples/main.py
@@ -6,7 +6,12 @@ from fastapi import FastAPI
 from starlette.middleware.cors import CORSMiddleware
 from starlette.responses import RedirectResponse
 from starlette.staticfiles import StaticFiles
-from starlette.status import HTTP_403_FORBIDDEN, HTTP_404_NOT_FOUND, HTTP_500_INTERNAL_SERVER_ERROR
+from starlette.status import (
+    HTTP_401_UNAUTHORIZED,
+    HTTP_403_FORBIDDEN,
+    HTTP_404_NOT_FOUND,
+    HTTP_500_INTERNAL_SERVER_ERROR,
+)
 from tortoise.contrib.fastapi import register_tortoise
 
 from examples import settings
@@ -18,6 +23,7 @@ from fastapi_admin.exceptions import (
     forbidden_error_exception,
     not_found_error_exception,
     server_error_exception,
+    unauthorized_error_exception,
 )
 
 
@@ -36,13 +42,13 @@ def create_app():
     admin_app.add_exception_handler(HTTP_500_INTERNAL_SERVER_ERROR, server_error_exception)
     admin_app.add_exception_handler(HTTP_404_NOT_FOUND, not_found_error_exception)
     admin_app.add_exception_handler(HTTP_403_FORBIDDEN, forbidden_error_exception)
+    admin_app.add_exception_handler(HTTP_401_UNAUTHORIZED, unauthorized_error_exception)
 
     @app.on_event("startup")
     async def startup():
-        redis = await aioredis.create_redis_pool(
-            f"redis://{settings.REDIS_HOST}:{settings.REDIS_PORT}",
-            db=settings.REDIS_DB,
-            password=settings.REDIS_PASSWORD,
+        redis = aioredis.from_url(
+            settings.REDIS_URL,
+            decode_responses=True,
             encoding="utf8",
         )
         await admin_app.configure(